// ManutenzioniClient/src/app/interfaces/IsPHomeCarico.ts
export interface IspHomeCarico {
    docid: number;
    società: string;
    tipo: string;
    fatturato: string;
    numOrd: string;
    startDate: string;
    customer: string;
    branch: string;
    cliente: string;
    indirizzo: string;
    cap: string;
    citta: string;
    prov: string;
    qta: number;
    nota: string;
    noterevcol: string;
    tipologiaAttivita: string;
    codTecnicoZona: string;
    pivaCf: string;
    tecnicoZona: string;
    id: number;
    position: number;
  }